format black

+++ meandian/meandian_include.py
@@ -1,5 +1,7 @@
 import numpy as np
 from numpy.core import numerictypes as nt
+from numba import njit
+
 
 def meandian(x, alpha=1.5, max_iter=100, tol=1e-10):
     # make sure the input is an array
@@ -12,16 +14,19 @@ def meandian(x, alpha=1.5, max_iter=100, tol=1e-10):
 
     # make sure the input is not an int or bool
     if issubclass(x.dtype.type, (nt.integer, nt.bool_)):
-        x = x.astype('f8')
+        x = x.astype("f8")
 
     meandian_values = np.zeros(alpha.shape)
     if np.sum(alpha < 1):
         meandian_values[alpha < 1] = brute_meandian2(x, alpha[alpha < 1])
 
     if np.sum(alpha >= 1):
-        meandian_values[alpha >= 1] = interval_meandian(x, alpha[alpha >= 1], max_iter=max_iter, tol=tol)
+        meandian_values[alpha >= 1] = interval_meandian(
+            x, alpha[alpha >= 1], max_iter=max_iter, tol=tol
+        )
 
     return meandian_values
+
 
 def get_meandian(x, alpha=1.5, max_iter=100, tol=1e-6, axis=None):
     single_alpha = False
@@ -37,8 +42,10 @@ def get_meandian(x, alpha=1.5, max_iter=100, tol=1e-6, axis=None):
         shape = y.shape
         yy = np.reshape(y, (-1, y.shape[-1]))
 
-        mus = np.array([interval_meandian_jit(xx, np.asarray(alpha), max_iter, tol) for xx in yy])
-        mus = np.reshape(mus, list(shape[:-1])+[-1])
+        mus = np.array(
+            [interval_meandian_jit(xx, np.asarray(alpha), max_iter, tol) for xx in yy]
+        )
+        mus = np.reshape(mus, list(shape[:-1]) + [-1])
         if single_alpha is True:
             return mus[..., 0]
         return mus
@@ -49,15 +56,16 @@ def get_meandian(x, alpha=1.5, max_iter=100, tol=1e-6, axis=None):
             return mus[0]
         return mus
 
-from numba import njit
+
 @njit
 def interval_meandian_jit(x, alpha=1.5, max_iter=100, tol=1e-6):
     results = []
 
     for a in alpha:
         if np.isinf(a):
-            results.append((np.max(x)+np.min(x))/2)
+            results.append((np.max(x) + np.min(x)) / 2)
         elif a < 1:
+
             def d(x, mu, n, axis=0):
                 x = x[:, None]
                 mu = mu[None, :]
@@ -69,6 +77,7 @@ def interval_meandian_jit(x, alpha=1.5, max_iter=100, tol=1e-6):
         elif i == 1:
             results.append(np.median(x))
         else:
+
             def dp(mu):
                 a_mu = x - mu
                 return np.sum(-np.abs(a_mu) ** (a - 1) * np.sign(a_mu))
@@ -86,6 +95,7 @@ def interval_meandian_jit(x, alpha=1.5, max_iter=100, tol=1e-6):
             results.append((interval_x[0] + interval_x[1]) / 2)
 
     return np.array(results)
+
 
 def interval_meandian(x, alpha=1.5, max_iter=100, tol=1e-10):
     if not isinstance(alpha, (tuple, list, np.ndarray)):
@@ -119,4 +129,3 @@ def brute_meandian2(x, alpha=1.5):
     v = d(x, x, alpha)
     i = np.argmin(v, axis=0)
     return x[i]
-
